move make_custom_builds_fn to its own file; use proper protocol for StdBuilds
